requirements.txt
ta==0.10.0
numpy>=1.20.0
pandas>=1.3.0
torch>=1.10.0
scikit-learn>=1.0.0
pyyaml>=6.0
matplotlib>=3.5.0